+++ CHANGE_LOG.md
@@ -1,3 +1,186 @@
+## [1.201.2](https://github.com/oaknational/Oak-Web-Application/compare/v1.201.1...v1.201.2) (2023-07-04)
+
+
+### Bug Fixes
+
+* **lesson-listing:** connect lesson listing to mv to expose all units ([9e77a2d](https://github.com/oaknational/Oak-Web-Application/commit/9e77a2d44f0ba9f6bb21e8dcc16b6a89f752e081))
+
+## [1.201.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.201.0...v1.201.1) (2023-07-03)
+
+
+### Bug Fixes
+
+* remove log statements from components ([d9e48a6](https://github.com/oaknational/Oak-Web-Application/commit/d9e48a6f705a9933e001093d3a7a037db93af679))
+
+# [1.201.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.200.0...v1.201.0) (2023-07-03)
+
+
+### Bug Fixes
+
+* exam board tiers only come from exam board programmes ([9afeb65](https://github.com/oaknational/Oak-Web-Application/commit/9afeb659f3f98f8483c3462a1fe181246b4df27a))
+
+
+### Features
+
+* add exam boards to programme page ([8c04e14](https://github.com/oaknational/Oak-Web-Application/commit/8c04e1454d2489baa4b511ec8df220801985360f))
+
+# [1.200.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.199.0...v1.200.0) (2023-06-30)
+
+
+### Features
+
+* programme listing 2023 api ([6989837](https://github.com/oaknational/Oak-Web-Application/commit/698983746cffee184d12828dde1fe4df3db481d5))
+
+# [1.199.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.198.1...v1.199.0) (2023-06-29)
+
+
+### Features
+
+* **unit_lisitng:** connect unit listing page to the 2023 hasura instance ([a576f64](https://github.com/oaknational/Oak-Web-Application/commit/a576f644248779fa9042c6dd01d6b2639b791c19))
+* **unit_lisitng:** prettier fix ([9eb05d8](https://github.com/oaknational/Oak-Web-Application/commit/9eb05d860061b195a8bda1fa593087b270555d3b))
+* **unit_lisitng:** remove console logs ([cef2dc3](https://github.com/oaknational/Oak-Web-Application/commit/cef2dc3053af2f04e9b5a7937e41d8ea2d8cab5a))
+* **unit_lisitng:** remove console logs ([b504ab2](https://github.com/oaknational/Oak-Web-Application/commit/b504ab26d71dc30fa1e4873fc3aa7a9f25b7688b))
+* **unit_lisitng:** unit lisitng fix ([5d4f0a3](https://github.com/oaknational/Oak-Web-Application/commit/5d4f0a3cb3d14ca69653b0fe9319ef6f606f8ce2))
+
+## [1.198.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.198.0...v1.198.1) (2023-06-29)
+
+# [1.198.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.197.0...v1.198.0) (2023-06-28)
+
+
+### Bug Fixes
+
+* broken tests ([135a6da](https://github.com/oaknational/Oak-Web-Application/commit/135a6da559a14edb5c4ced9f569c2d6e3149d156))
+* lesson listing zod error test ([1e083fc](https://github.com/oaknational/Oak-Web-Application/commit/1e083fc6342d215a79a9f59dd2b1e7e99723e97b))
+* storybook ([4f35a15](https://github.com/oaknational/Oak-Web-Application/commit/4f35a151aa1f80f836cd8142ae3b41cd1deac01e))
+
+
+### Features
+
+* subject listing 2023 ([c323617](https://github.com/oaknational/Oak-Web-Application/commit/c3236177a939c97130c4e27718772456d2ee5022))
+
+# [1.197.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.196.1...v1.197.0) (2023-06-28)
+
+
+### Features
+
+* getPageProps() function with error handling and isr ([5d68f2f](https://github.com/oaknational/Oak-Web-Application/commit/5d68f2f22c15ad7578264d8beafd700d862dcf27))
+* isr error handling ([9b29ca6](https://github.com/oaknational/Oak-Web-Application/commit/9b29ca6bf8f50dfd2ec133cde378a0fc739b3ea3))
+
+## [1.196.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.196.0...v1.196.1) (2023-06-27)
+
+
+### Bug Fixes
+
+* change copy on newsletter form ([84c82e4](https://github.com/oaknational/Oak-Web-Application/commit/84c82e4a4d1fb8baee1a5374cae70b3b08582747))
+
+# [1.196.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.195.0...v1.196.0) (2023-06-27)
+
+
+### Bug Fixes
+
+* add missing npm ci step ([ced4c29](https://github.com/oaknational/Oak-Web-Application/commit/ced4c29fc749a21bcf382aaf513aa975452be898))
+* use forked version of updater action with explicit prNumber passed ([e59f7c0](https://github.com/oaknational/Oak-Web-Application/commit/e59f7c0fe6e05ee29c581e782c4edee17d6239a5))
+
+
+### Features
+
+* automatically add the deployment URL to PR descriptions ([763372a](https://github.com/oaknational/Oak-Web-Application/commit/763372a89f2850b1f27a6ddc589e124a7e5eb3b9))
+* swap {cloud link} for {owa_deployment_url} in PR template ([0dac2d9](https://github.com/oaknational/Oak-Web-Application/commit/0dac2d99d1c2967c7f7d778d737320105c6b8a9b))
+
+# [1.195.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.194.0...v1.195.0) (2023-06-26)
+
+
+### Features
+
+* **downloads:** connect downloads page to 2023 hasura end point ([3feadfb](https://github.com/oaknational/Oak-Web-Application/commit/3feadfb04acada5d58c3429f09af18a2f89d06e6))
+* **downloads:** fix type error on fixture ([e8bdb09](https://github.com/oaknational/Oak-Web-Application/commit/e8bdb0921b8d0d11c65e534a74e8967090416331))
+* **downloads:** prettier fix ([c112c6d](https://github.com/oaknational/Oak-Web-Application/commit/c112c6d934e85ca3d5644092a31d012010cdac65))
+* **downloads:** prettier fix ([4b4ed80](https://github.com/oaknational/Oak-Web-Application/commit/4b4ed806fda60b051e757492bc9a8e37d3e081c6))
+* **downloads:** remove duplicate code ([72b81d2](https://github.com/oaknational/Oak-Web-Application/commit/72b81d2999f80d8b2bd085976604cfe870abd781))
+
+# [1.194.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.193.0...v1.194.0) (2023-06-26)
+
+
+### Bug Fixes
+
+* only show preview share button when secret param is present ([2a190cf](https://github.com/oaknational/Oak-Web-Application/commit/2a190cf9a96fb79661f50b814dc24f58a988db55))
+
+
+### Features
+
+* allow copying the current page's preview URL from PreviewControls ([060c6f2](https://github.com/oaknational/Oak-Web-Application/commit/060c6f2f56ba53f51e24e82fb7e376fe35d5eb5f))
+
+# [1.193.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.192.0...v1.193.0) (2023-06-23)
+
+
+### Features
+
+* **homepage:** connect to 2023 data ([5951297](https://github.com/oaknational/Oak-Web-Application/commit/5951297e0128663b83c5bd0f01912fbec8a248be))
+* **homepage:** remove console log ([1a44df1](https://github.com/oaknational/Oak-Web-Application/commit/1a44df1100d5a8d2f984b871a486c88cc9c45f3e))
+
+# [1.192.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.191.0...v1.192.0) (2023-06-22)
+
+
+### Features
+
+* add mv-lesson-overview to curriculum api ([b3d4f4f](https://github.com/oaknational/Oak-Web-Application/commit/b3d4f4faec646d9f10bbe17204937d087653c5aa))
+
+# [1.191.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.190.0...v1.191.0) (2023-06-22)
+
+
+### Features
+
+* **search:** cleanup code and test suite ([01aa1a4](https://github.com/oaknational/Oak-Web-Application/commit/01aa1a4f3c66be3bd12c0e59903fedd9e089c0cb))
+* **search:** cleanup code and test suite ([9a06722](https://github.com/oaknational/Oak-Web-Application/commit/9a06722939ddd19eac4f3b94022d1064261bba27))
+* **search:** connect 2023 to hasura for search page ([0e65456](https://github.com/oaknational/Oak-Web-Application/commit/0e654562f5f115ebebc3692a0506b45d47f16de3))
+* **search:** prettier fix ([547144f](https://github.com/oaknational/Oak-Web-Application/commit/547144fdb491926e060291c1111f7986b68c308a))
+
+# [1.190.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.189.0...v1.190.0) (2023-06-20)
+
+
+### Bug Fixes
+
+* add missing CTA to landing page Text & Media blocks ([3a240b4](https://github.com/oaknational/Oak-Web-Application/commit/3a240b43dd62d044300e6bca4dc944182cfde31d))
+* add spacing above blog post CTAs ([b257ec5](https://github.com/oaknational/Oak-Web-Application/commit/b257ec54fe7d0a23d9c8763a43edf4201cfac457))
+* left-align quote attribution on blog posts ([bd5100e](https://github.com/oaknational/Oak-Web-Application/commit/bd5100eda245f681c4b04dfe7ef57642a7244192))
+* use renamed getLinkHref function in LandingPageTextAndMedia ([89916ed](https://github.com/oaknational/Oak-Web-Application/commit/89916edaad0ab3dc5f1dfe2a87f25265cd0c9999))
+
+
+### Features
+
+* add form block to blog posts ([740bcc3](https://github.com/oaknational/Oak-Web-Application/commit/740bcc3cb8a722d4f9162924c1b53e7d5cb84d9e))
+
+# [1.189.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.188.2...v1.189.0) (2023-06-19)
+
+
+### Bug Fixes
+
+* don't render footnotes footer when none exist ([a692cf3](https://github.com/oaknational/Oak-Web-Application/commit/a692cf3a71e3fbeee0b48baaaddf2bfe3cf78839))
+* ensure footnote backlinks scroll to correct position ([5c6aedf](https://github.com/oaknational/Oak-Web-Application/commit/5c6aedfa47699e49968818cdc3d7ac36a187ef61))
+* prevent mangling of footnote anchor hrefs by OakLink/Next's Link ([2a82c57](https://github.com/oaknational/Oak-Web-Application/commit/2a82c5727fc2125ad9e7faad4a88890337a0e83a))
+
+
+### Features
+
+* add ability to mark up blog posts with footnotes ([8cb377f](https://github.com/oaknational/Oak-Web-Application/commit/8cb377f25651132684a71f449e27f63c5318228c))
+* tidy footnote styling, add tests ([849faa8](https://github.com/oaknational/Oak-Web-Application/commit/849faa8a35f5199236258df5eb39aef18f8dc158))
+
+## [1.188.2](https://github.com/oaknational/Oak-Web-Application/compare/v1.188.1...v1.188.2) (2023-06-15)
+
+## [1.188.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.188.0...v1.188.1) (2023-06-15)
+
+
+### Bug Fixes
+
+* **E2E Browser:** allow tests to run against prod ([f5a685d](https://github.com/oaknational/Oak-Web-Application/commit/f5a685db19124e5a764d78b1733f9f71406cd2e8))
+
+# [1.188.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.187.1...v1.188.0) (2023-06-14)
+
+
+### Features
+
+* spike curriculum api 2023 depending on url param ([acd0533](https://github.com/oaknational/Oak-Web-Application/commit/acd0533aa90befcd1cd7fdcdc9bb464640453ff7))
+
 ## [1.187.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.187.0...v1.187.1) (2023-06-13)